musician page is almost complete
