+++ client/public/css/main.css
@@ -6,7 +6,7 @@ html {
     margin-top: 100;
 }
 
-.calendar { margin-left: -1% }
+.calendar { margin-left: -1%; }
 
 .cld {
     height: 50px;
